+++ test/auth/oidc_test.go
@@ -18,54 +18,3 @@ limitations under the License.
 */
 
 package auth
-
-import (
-	"testing"
-	"time"
-
-	"knative.dev/reconciler-test/pkg/eventshub"
-
-	"knative.dev/pkg/system"
-	"knative.dev/reconciler-test/pkg/environment"
-	"knative.dev/reconciler-test/pkg/feature"
-	"knative.dev/reconciler-test/pkg/k8s"
-	"knative.dev/reconciler-test/pkg/knative"
-
-	"knative.dev/eventing/test/auth/features/oidc"
-	brokerfeatures "knative.dev/eventing/test/rekt/features/broker"
-	"knative.dev/eventing/test/rekt/resources/broker"
-)
-
-func TestBrokerSupportsOIDC(t *testing.T) {
-	t.Parallel()
-
-	ctx, env := global.Environment(
-		knative.WithKnativeNamespace(system.Namespace()),
-		knative.WithLoggingConfig,
-		knative.WithTracingConfig,
-		k8s.WithEventListener,
-		environment.Managed(t),
-		environment.WithPollTimings(4*time.Second, 12*time.Minute),
-		eventshub.WithTLS(t),
-	)
-
-	name := feature.MakeRandomK8sName("broker")
-	env.Prerequisite(ctx, t, brokerfeatures.GoesReady(name, broker.WithEnvConfig()...))
-
-	env.TestSet(ctx, t, oidc.AddressableOIDCConformance(broker.GVR(), "Broker", name, env.Namespace()))
-}
-
-func TestBrokerSendsEventsWithOIDCSupport(t *testing.T) {
-	t.Parallel()
-
-	ctx, env := global.Environment(
-		knative.WithKnativeNamespace(system.Namespace()),
-		knative.WithLoggingConfig,
-		knative.WithTracingConfig,
-		k8s.WithEventListener,
-		environment.Managed(t),
-		eventshub.WithTLS(t),
-	)
-
-	env.TestSet(ctx, t, oidc.BrokerSendEventWithOIDC())
-}
